Creating subdirectories for config and utils
